chore: next docs navaigation cleanup
