+++ next-docs/components/sidebar/navigation.ts
@@ -3,11 +3,11 @@ const navigation = [
   { name: 'Vision', href: '/vision' },
   { name: 'Getting Started', href: '/getting-started' },
   { name: 'How to contribute', href: '/how-to-contribute' },
+  // { name: 'Legacy Colours Palette', href: '/colours-legacy' },
   { name: 'Colours Palette', href: '/colours' },
-  { name: 'Colours Palette New', href: '/colours-new' },
   { name: 'Typography', href: '/typography' },
+  // { name: 'Legacy Tokens', href: '/tokens-legacy' },
   { name: 'Tokens', href: '/tokens' },
-  { name: 'Tokens New', href: '/tokens-new' },
   { name: 'Icons', href: '/icons' },
   {
     name: 'Components',
@@ -39,9 +39,9 @@ const navigation = [
       {
         name: 'Dialog',
         children: [
-          { name: 'Dialog', href: '/components/dialog/dialog' }, // TODO code
-          { name: 'Dialog Content', href: '/components/dialog/dialogContent' }, // TODO code
-          { name: 'Dialog Overlay', href: '/components/dialog/dialogOverlay' }, // TODO code
+          { name: 'Dialog', href: '/components/dialog/dialog' },
+          { name: 'Dialog Content', href: '/components/dialog/dialogContent' },
+          { name: 'Dialog Overlay', href: '/components/dialog/dialogOverlay' },
           { name: 'Dialog Header', href: '/components/dialog/header' },
           { name: 'Dialog Footer', href: '/components/dialog/footer' },
         ],
@@ -49,7 +49,7 @@ const navigation = [
       { name: 'Label', href: '/components/label' },
       { name: 'List Item', href: '/components/listItem' },
       { name: 'Loader', href: '/components/loader' },
-      { name: 'Pagination', href: '/components/pagination' }, // TODO
+      { name: 'Pagination', href: '/components/pagination' },
       {
         name: 'Progress',
         children: [
